Implement additional permutations of AclustsCCA

Implement an iterative penalized least squares approach to sparse canonical correlation analysis (SparseCCA) with various penalty functions.

AclustsCCA.cont(obj, X, Y, maxnum = NULL, maxB = 10000)

Arguments

obj

A result of AclustsCCA function.

X

\(n\) by \(p\) exposure data matrix, where \(n\) is sample size and \(p\) is number of exposures.

Y

\(n\) by \(q\) outcome data matrix, where \(n\) is sample size and \(q\) is number of outcomes.

maxnum

A maximal total number of permutations across all the clusters.

maxB

A maximal number of permutations for a single cluster.

Value

The function returns a list of 6 objects according to the following order:

clusters.list : A list of clusters with CpG sites obtained using A-clustering, each item is a cluster that contains a set of probes. If A-clustering is not implemented inside AclustsCCA, return NA.

ALPHA.observed : A list of estimated canonical vector of length \(p\) corresponding to the exposure data \(X\) for each cluster.

BETA.observed : A list of estimated canonical vector of length \(q\) corresponding to the outcome data \(Y\) for each cluster.

cancors.observed : A vector of estimated canonical correlation for each cluster.

permutation.result : A mmctest object that contains permutation results.

settings : A settings used for the analysis.

Examples