Sample Annotation Data for Analysis

A data.table that contains annotation for the probes on chromosome 7 in the Illumina 450K chip, sorted by chromosome and chromosomal coordinates of the CpG. created using the R package IlluminaHumanMethylation450kanno.ilmn12.hg19 from Bioconductor.

annot

Format

A data.table with 6 column, which are:

IlmnID

a character vector of IlmnID

CHR

a numeric vector of chromosome containing the CpG

Coordinate_37

a numeric vector of chromosomal coordinates of the CpG

Islands_Name

a character vector of chromosomal coordinates of the CpG Island

Relation_to_Island

a character vector of the location of the CpG relative to the CpG island

UCSC_RefGene_Name

a character vector of Target gene name(s)