Sample Data for AclustsCCA R Package

Generates synthetic data with p=20 exposures and q=100 DNA methylation probe data in chromosome 7. A correlation structure of exposures are defined as a block diagonal where each block is size of five with the first order autoregressive correlation structure of correlation 0.7. A correlation structure of outcomes are defined as a block diagonal where each block is size of its clusters defined by bumphunter with the first order autoregressive correlation structure of correlation 0.9. The data is generated from multivariate normal distribution.

A five exposures (exposures 1,2,3,9, and 10) are associated with a five DMR regions (each of size 2,3,4,7,and 12).

sample.data

Format

A list of length 10, which are:

DATA.X

A \(n=500\) by \(p\) matrix where \(n\) is a number of subjects and \(p\) is a number of exposures.

DATA.Y

A \(n=500\) by \(q\) matrix where \(n\) is a number of subjects and \(q\) is a number of CpG sites.

DATA.Z

A \(n=500\) by 2 matrix where \(n\) is a number of subjects and 2 is a number of confounders.

TRUE.CpGs

A vector of 28 true CpG sites that are associated with exposures

TRUE.Exposures

A vector of 5 true exposures that are associated with outcome

TRUE.ALPHA

A vector of length \(p\) which represents true loading vector \(alpha\)

TRUE.BETA

A vector of length \(q\) which represents true loading vector \(beta\)

SIGMA.XX

A \(p\) by \(p\) exposure correlation matrix used to generated data

SIGMA.YY

A \(q\) by \(q\) outcome correlation matrix used to generated data

SIGMA.XY

A \(p\) by \(q\) correlation matrix of exposure and outcome used to generated data

clusters.list

A list of 9 clusters with CpG sites obtained using A-clustering, each item is a cluster. Cluster 2,5,6,7,9 are the true DMRs.