Summarize the A-clustering result

Implement an iterative penalized least squares approach to sparse canonical correlation analysis (SparseCCA) with various penalty functions

INPUT

# S3 method for Aclustering
summary(clusters.list, annot, digits = 2)

Arguments

clusters.list

: A list of clusters with CpG sites, each item is a cluster that contains a set of probes

annot

: A preloaded annotation file that includes columns the below name. For more information, https://support.illumina.com/content/dam/illumina-support/documents/downloads/productfiles/methylationepic/infinium-methylationepic-manifest-column-headings.pdf.

IlmnID: The IlmnID

Coordinate_37: Chromosomal coordinates of the CpG

Islands_Name: Chromosomal coordinates of the CpG Island

Relation_to_Island: The location of the CpG relative to the CpG island

UCSC_RefGene_Name: Target gene name(s)

digits

: Show rounded values to the specified number of decimal places

Value

The function returns a table of summary characteristics:

Total number of sites used for the analysis

Number of singletons (a cluster with one CpG site)

Number of clusters (a cluster with at least two sites)

Number of CpG sites that are clustered

Number of CpG sites in cluster (min, median, max)

Base-pair distance between extremes (min, median, max)

(1) Clusters associated with a single refernece gene

(2) Clusters associated with a single CpG Island

(3-1) Clusters associated with a CpG resort : at least one resort regions

(3-2) Clusters associated with a CpG resort : at least two resort regions

(3-3) Clusters associated with a CpG resort : one resort regions

(3-4) Clusters associated with a CpG resort : two resort regions

(3-5) Clusters associated with a CpG resort : three resort regions

Examples

library("IlluminaHumanMethylation450kanno.ilmn12.hg19")
#> Loading required package: minfi
#> Loading required package: BiocGenerics
#> Loading required package: parallel
#> 
#> Attaching package: 'BiocGenerics'
#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB
#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from 'package:base':
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: S4Vectors
#> 
#> Attaching package: 'S4Vectors'
#> The following objects are masked from 'package:base':
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> 
#> Attaching package: 'IRanges'
#> The following object is masked from 'package:grDevices':
#> 
#>     windows
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: 'MatrixGenerics'
#> The following objects are masked from 'package:matrixStats':
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: 'Biobase'
#> The following object is masked from 'package:MatrixGenerics':
#> 
#>     rowMedians
#> The following objects are masked from 'package:matrixStats':
#> 
#>     anyMissing, rowMedians
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: 'Biostrings'
#> The following object is masked from 'package:base':
#> 
#>     strsplit
#> Loading required package: bumphunter
#> Loading required package: foreach
#> Loading required package: iterators
#> Loading required package: locfit
#> Warning: package 'locfit' was built under R version 4.1.2
#> locfit 1.5-9.4 	 2020-03-24
#> Setting options('download.file.method.GEOquery'='auto')
#> Setting options('GEOquery.inmemory.gpl'=FALSE)
annot <- minfi::getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)
annot <- data.table(data.frame(annot))
#> Error in data.table(data.frame(annot)): could not find function "data.table"
annot <- annot[!chr%in%c("chrY","chrX"),]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': error in evaluating the argument 'x' in selecting a method for function '%in%': object 'chr' not found
annot <- annot[,list("IlmnID"=Name,"Coordinate_37"=pos,"CHR"=as.numeric(gsub("chr", "", chr)),Islands_Name,Relation_to_Island,UCSC_RefGene_Name)]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'j' in selecting a method for function '[': object 'Name' not found
annot <- annot[order(CHR,Coordinate_37),]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': object 'CHR' not found

clusters.list <- assign.to.clusters(betas.7, annot)
#> Error in assign.to.clusters(betas.7, annot): could not find function "assign.to.clusters"
summary.Aclustering(clusters.list=clusters.list,annot=annot,digits=2)
#> Error in summary.Aclustering(clusters.list = clusters.list, annot = annot,     digits = 2): could not find function "summary.Aclustering"