Create a table that summarizes A-clustering result

This function creates a table that summarizes A-clustering result.

summary_Aclustering(clusters.list, annot)

Arguments

clusters.list

A list of clusters with CpG sites, each item is a cluster that contains a set of probes

annot

A preloaded annotation file that includes columns the below name. For more information, IlluminaHumanMethylation450kanno.ilmn12.hg19.

IlmnID: a character vector of IlmnID

CHR: a numeric vector of chromosome containing the CpG

Coordinate_37: a numeric vector of chromosomal coordinates of the CpG

Islands_Name: a character vector of chromosomal coordinates of the CpG Island

Relation_to_Island: a character vector of the location of the CpG relative to the CpG island

UCSC_RefGene_Name: a character vector of Target gene name(s)

Value

The function returns a table of summary characteristics:

Total number of sites used for the analysis

Number of singletons (a cluster with one CpG site)

Number of clusters (a cluster with at least two sites)

Number of CpG sites that are clustered

Number of CpG sites in cluster (min, median, max)

Base-pair distance between extremes (min, median, max)

(1) Clusters associated with a single refernece gene

(2) Clusters associated with a single CpG Island

(3) Number of resort regions each cluster is associated with (min, median, max)

(3-1) Clusters associated with a CpG resort : at least one resort regions

(3-2) Clusters associated with a CpG resort : at least two resort regions

Examples

data(annot)
data(sample.data)
DATA.Y <- sample.data$DATA.Y # row: subjects (n), column: CpG sites (q)

all.clusters.list <- Aclust::assign.to.clusters(betas = t(DATA.Y),annot = annot, dist.type = "spearman", method = "average", thresh.dist = 0.2,max.dist = 1000,bp.thresh.dist = 999)
#> Error in IlmnID %in% rownames(betas): object 'IlmnID' not found
summary_Aclustering(clusters.list=all.clusters.list,annot=annot)
#> Error in lapply(seq_along(clusters.list), function(i) {    data.frame(IlmnID = clusters.list[[i]], ClustIdx = i)}): object 'all.clusters.list' not found