Outputs summary of AclustsCCA result

This function creates summary of AclustsCCA result table that summarizes A-clustering result.

summary_AclustsCCA(obj, annot, n.top)

Arguments

obj

Output of AclustsCCA

annot

A preloaded annotation file that includes columns the below name. For more information, IlluminaHumanMethylation450kanno.ilmn12.hg19.

IlmnID: a character vector of IlmnID

CHR: a numeric vector of chromosome containing the CpG

Coordinate_37: a numeric vector of chromosomal coordinates of the CpG

Islands_Name: a character vector of chromosomal coordinates of the CpG Island

Relation_to_Island: a character vector of the location of the CpG relative to the CpG island

UCSC_RefGene_Name: a character vector of Target gene name(s)

n.top

A number of top clusters with smallest p.value to show

Value

The function returns a table of summary characteristics for each cluster:

ClustIdx: Index number of cluster corresponding to clusters.list

CHR: a numeric value of chromosome containing the CpG

UCSC_RefGene_Name: a list of target gene name(s)

Islands_Name: a list of chromosomal coordinates of the CpG Island

Significant: whether or not that cluster is significant

Cancors: a numeric value of canonical correlation

N.Probes: a number of probes in that cluster

DMR.length: a length of DMR in base pairs

Num.Exceed: a number of exceedness in permutation test

Num.Perm: a number of permutation

P.value: P.value corresponding to this cluster

CpGs: CpG sites with non-zero loadings

Exposures: Exposures with non-zero loadings

CpGs.Loading: A loading vector of CpGs sites with non-zero loadings

Exposures.Loading: A loading vector of exposures with non-zero loadings